chore: refactoring MainScreenViewModel.kt by taking logic to managers
