+++ app/src/main/kotlin/model/command/concrete/Commands.kt
@@ -118,9 +118,9 @@ class Commands<E : Comparable<E>, V : Comparable<V>>(
             command.parameters["strategy"] ?: return Result.Error(CommandErrors.MissingParameters("Strategy is required"))
         if (viewModel != null) {
             val strategy =
-                viewModel.getLayoutStrategyByString(strategyAsString)
+                viewModel.graphManager.getLayoutStrategyByString(strategyAsString)
                     ?: return Result.Error(CommandErrors.UnknownLayoutStrategy(strategyAsString))
-            viewModel.updateLayoutStrategy(strategy)
+            viewModel.graphManager.updateLayoutStrategy(strategy)
             return Result.Success("Layout strategy updated. New strategy: $strategyAsString")
         } else {
             return Result.Error(CommandErrors.ViewmodelNotFounded())
