+++ app/src/main/kotlin/view/appInterface/layout/TitleBar.kt
@@ -158,7 +158,7 @@ fun <E : Comparable<E>, V : Comparable<V>> TitleBar(
                         ) {
                             DropdownMenuItem(
                                 onClick = {
-                                    val loadResult = viewModel.openGraphFile()
+                                    val loadResult = viewModel.graphManager.openGraphFile()
                                     warnings =
                                         when (loadResult) {
                                             is Result.Success -> loadResult.data
@@ -191,7 +191,7 @@ fun <E : Comparable<E>, V : Comparable<V>> TitleBar(
                             }
                             DropdownMenuItem(
                                 onClick = {
-                                    val openResult = viewModel.openGraphRepository()
+                                    val openResult = viewModel.graphManager.openGraphRepository()
                                     when (openResult) {
                                         is Result.Error ->
                                             warnings =
@@ -230,10 +230,10 @@ fun <E : Comparable<E>, V : Comparable<V>> TitleBar(
                     }
                     DropdownMenuItem(
                         onClick = {
-                            if (viewModel.graphPath == null) {
+                            if (viewModel.graphManager.graphPath == null) {
                                 showSaveAsDialog = true
                             } else {
-                                val saveResult = viewModel.saveGraph()
+                                val saveResult = viewModel.graphManager.saveGraph()
                                 if (saveResult is Result.Error) {
                                     userNotificationMessage =
                                         LocalizationFormatter.getErrorMessage(
@@ -247,14 +247,14 @@ fun <E : Comparable<E>, V : Comparable<V>> TitleBar(
                             }
                             showFileMenu = false
                         },
-                        enabled = viewModel.isGraphActive,
+                        enabled = viewModel.graphManager.isGraphActive,
                     ) {
                         Row(verticalAlignment = Alignment.CenterVertically) {
                             Icon(
                                 imageVector = Icons.Filled.Save,
                                 contentDescription = LocalizationManager.states.ui.titleBarFileSaveIconDescription.value,
                                 tint =
-                                    if (viewModel.isGraphActive) {
+                                    if (viewModel.graphManager.isGraphActive) {
                                         MaterialTheme.colors.primary
                                     } else {
                                         MaterialTheme.colors.onSurface.copy(
@@ -267,7 +267,7 @@ fun <E : Comparable<E>, V : Comparable<V>> TitleBar(
                                 text = LocalizationManager.states.ui.titleBarFileSaveButton.value,
                                 style = MaterialTheme.typography.button,
                                 color =
-                                    if (viewModel.isGraphActive) {
+                                    if (viewModel.graphManager.isGraphActive) {
                                         MaterialTheme.colors.onSurface
                                     } else {
                                         MaterialTheme.colors.onSurface
@@ -283,14 +283,14 @@ fun <E : Comparable<E>, V : Comparable<V>> TitleBar(
                             showFileMenu = false
                             showSaveAsDialog = true
                         },
-                        enabled = viewModel.isGraphActive,
+                        enabled = viewModel.graphManager.isGraphActive,
                     ) {
                         Row(verticalAlignment = Alignment.CenterVertically) {
                             Icon(
                                 imageVector = Icons.Filled.SaveAs,
                                 contentDescription = LocalizationManager.states.ui.titleBarFileSaveAsIconDescription.value,
                                 tint =
-                                    if (viewModel.isGraphActive) {
+                                    if (viewModel.graphManager.isGraphActive) {
                                         MaterialTheme.colors.primary
                                     } else {
                                         MaterialTheme.colors.onSurface.copy(
@@ -303,7 +303,7 @@ fun <E : Comparable<E>, V : Comparable<V>> TitleBar(
                                 text = LocalizationManager.states.ui.titleBarFileSaveAsButton.value,
                                 style = MaterialTheme.typography.button,
                                 color =
-                                    if (viewModel.isGraphActive) {
+                                    if (viewModel.graphManager.isGraphActive) {
                                         MaterialTheme.colors.onSurface
                                     } else {
                                         MaterialTheme.colors.onSurface
@@ -386,8 +386,8 @@ fun <E : Comparable<E>, V : Comparable<V>> TitleBar(
             onDismiss = { showOpenRepositoryDialog = false },
             graphs = graphs,
             onGraphSelected = { graphId ->
-                viewModel.graphId = graphId
-                val loadResult = viewModel.loadGraphFromFile(selectedRepositoryFile)
+                viewModel.graphManager.graphId = graphId
+                val loadResult = viewModel.graphManager.loadGraphFromFile(selectedRepositoryFile)
                 if (loadResult is Result.Error) {
                     warnings =
                         listOf(
@@ -407,11 +407,11 @@ fun <E : Comparable<E>, V : Comparable<V>> TitleBar(
 
     if (showSaveAsDialog) {
         SaveGraphAs(
-            graphName = viewModel.graphName,
+            graphName = viewModel.graphManager.graphName,
             onDismiss = { showSaveAsDialog = false },
             onSave = { savedGraphDetails ->
                 val saveResult =
-                    viewModel.saveGraph(
+                    viewModel.graphManager.saveGraph(
                         fileName = savedGraphDetails.fileName,
                         directoryPath = savedGraphDetails.directoryPath,
                         fileFormat = savedGraphDetails.fileFormat,
@@ -454,11 +454,11 @@ fun <E : Comparable<E>, V : Comparable<V>> TitleBar(
         NewGraph(
             onDismiss = { showNewGraphDialog = false },
             onCreate = { newGraphData ->
-                viewModel.graphName = newGraphData.graphName
-                viewModel.graphAuthor = "None"
-                viewModel.graphPath = null
-                viewModel.graphFormat = null
-                viewModel.updateGraph(newGraphData.graph)
+                viewModel.graphManager.graphName = newGraphData.graphName
+                viewModel.graphManager.graphAuthor = "None"
+                viewModel.graphManager.graphPath = null
+                viewModel.graphManager.graphFormat = null
+                viewModel.graphManager.updateGraph(newGraphData.graph)
             },
         )
     }
